Move JSON code below /codec
To prevent package import cycles, the interface
`EncodeDecoder` is also moved into `codec` where it fits naturally.

Fixes #19
