+++ codec/json/json_test.go
@@ -1,0 +1,11 @@
+package json
+
+import (
+	"testing"
+
+	"github.com/asdine/storm/codec"
+)
+
+func TestJSON(t *testing.T) {
+	codec.RountripTester(t, Codec)
+}
